Move dir cap to vkdx11_interop
