+++ CMakeLists.txt
@@ -41,6 +41,6 @@ macro(_copy_file_to_target target thefile)
     )
 endmacro()
 
-add_subdirectory(capi)
+add_subdirectory(vkdx11_interop)
 add_subdirectory(rprpp)
 add_subdirectory(apps)
