+++ rprpp/vk_helper.h
@@ -1,8 +1,7 @@
 #pragma once
 
-#include "capi/rprpp.h"
+#include "vkdx11_interop/rprpp.h"
 #include "vk.h"
-#include <optional>
 
 namespace vk::helper {
