+++ vkdx11_interop/version.rc.in
