+++ vkdx11_interop/version_info.h.in
